requirements.txt
boto3==1.12.43
botocore==1.15.43
docutils==0.15.2
jmespath==0.9.5
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
boto3==1.12.43
botocore==1.15.43
docutils==0.15.2
jmespath==0.9.5
Pillow==7.1.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
boto3==1.12.43
botocore==1.15.43
docutils==0.15.2
jmespath==0.9.5
Pillow==7.1.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
appnope==0.1.0
backcall==0.1.0
boto3==1.12.43
botocore==1.15.43
decorator==4.4.2
docutils==0.15.2
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
jmespath==0.9.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
